Consolidate toolkit into unified GUI
